[Refactor] modalAnimation 삭제 반영 #387
